Move caching/invalidation to RecordedRcon for affected commands
